Refactor element picker menu to own file
